extract pigeon async callback into common method
